+++ forte2/mp/__init__.py
@@ -1,1 +1,3 @@
-from .dfmp2 import RHFMP2, ROHFMP2, UHFMP2
+from .rmp2 import RMP2
+from .romp2 import ROMP2
+from .ump2 import UMP2
